migrated "bot send" subcommand to "tgcli.cli.bot.send" package
