+++ .vscode/settings.json
@@ -1,5 +1,16 @@
 {
-    "python.linting.enabled": false,
-    "python.linting.pep8Enabled": true,
-    "python.linting.pylintEnabled": false
+    "python.linting.enabled": true,
+    "python.linting.pep8Enabled": false,
+    "python.linting.pylintEnabled": true,
+    "files.exclude": {
+        "**/.git": true,
+        "**/.svn": true,
+        "**/.hg": true,
+        "**/CVS": true,
+        "**/.DS_Store": true,
+        "**/__pycache__": true,
+        ".venv": true,
+        ".vscode": true
+    },
+    "explorerExclude.backup": null
 }
